Check if driver already have a ride
New updates in ride requests table can occur while the driver don't see the passengers selection screen. For that reason, the requests must be retrieved while listening to new updates when the page is reload. The maps: approved and requests were added in order to reduce the possibility of race conditions