Reapplying validation if ride already exists in RideCreationScreen
In case some ride with the same id still exists by the moment of ride creation. This scenario may only occur when two users are logged with the same user and both decide to create the ride at the same time. In other words, this is a very rare scenario.